rm -r calls deltree

+++ shell/cmds/rm.js
@@ -1,8 +1,21 @@
+const exec = require('./exec')
 const {enc} = require('../utils/enc')
 
 module.exports = async (web3, kernel, cmd, args) => {
+  if (!args.length) {
+    return console.log('Need an argument.')
+  }
+  let deltree
+  if (args[0] === '-r') {
+    args.shift()
+    deltree = true
+  }
   await args.reduce(async (promise, x) => {
     await promise
-    await kernel.unlink(enc(x))
+    if (deltree) {
+      await exec(web3, kernel, '/bin/deltree', [x])
+    } else {
+      await kernel.unlink(enc(x))
+    }
   }, Promise.resolve())
 }
